// app/(main)/karyawan/Dashboard/components/types.ts
export interface RingkasanStats {
    totalHadir: number;
    totalTidakHadir: number;
    sisaCuti: number;
    totalJatahCuti: number;
}

export interface EmployeeDetails {
    officeName: string;
    officeAddress: string;
    division: string;
    position: string;
    officeLat: number;
    officeLon: number;
    maxDistance: number;
}